feat(Audio): move SampleLoader to Mixer and make filepaths JAR-safe
